// src/containers/wikipediaSearch/enums/Resource.ts
/**
 * Resource enum
 */
export enum Resource {
	/** Search phrase input placeholder text */
	SearchPhraseInputPlaceholder = "Search phrase",
	/** Replace with input  placeholder text */
	ReplaceWithInputPlaceholder = "Replace with",
	/** Search button text */
	SearchButton = "Search",
	/** Replace button text */
	ReplaceButton = "Replace",
	/** Replace all button text */
	ReplaceAllButton = "Replace all",
	/** Search result label text */
	SearchResultLabel = "Search result:",
	/** Found count label text */
	FoundCountLabel = "Found items: ",
	/** Visible count label text */
	VisibleCountLabel = "Visible items: ",
	/** Title table header text */
	TitleTableHeader = "Title",
	/** Snipet table header text */
	SnippetTableHeader = "Snippet",
	/** Loading text */
	Loading = "Loading...",
	/** Search error text */
	SearchError = "Search error!",
}